Fix: Fixed the issue where GenerateTodayReport does not return today assessments
This is an issue of DateTime.Now vs. DateTime.Now. For example, the assessment happens on 10 AM and the second assessment happens on 11AM on the same day. These two assessments will not be recognize as assessment today because the method will only get the assessments on the specific time not on the specific day.