Fix: Fixed the issue where GenerateTodayReport does not return today logs
This is an issue of DateTime.Now vs. DateTime.Now. For example, the first log happens on 10 AM and the second log happens on 11AM of the same day. These two logs will not be recognized as a log today because the method will only get the logs on the specific time the method was called not on the day that the log was created.