Fix: Fixed the issue where GenerateTodayReport does not return today patients
This is an issue of DateTime.Now vs. DateTime.Now. For example, the first patient was created on 10 AM and the second patient was created on 11AM of the same day. These two patients will not be recognized as a patient today because the method will only get the patients on the specific time the method was called not on the day that the patient was created.