Fix: Fixed the issue where GenerateTodayReport does not return today users
This is an issue of DateTime.Now vs. DateTime.Now. For example, the first user was created on 10 AM and the second user was created on 11AM of the same day. These two users will not be recognized as a user today because the method will only get the users on the specific time the method was called not on the day that the user was created.